Extract import model run method to use case
